Change LTR's operand to be a u16
This makes it pick the correct register size.

+++ lib/x86/processor.h
@@ -191,7 +191,7 @@ static inline u16 sldt(void)
     return val;
 }
 
-static inline void ltr(unsigned val)
+static inline void ltr(u16 val)
 {
     asm volatile ("ltr %0" : : "rm"(val));
 }
